* 1. When the “Download Asset ID” and “Download Asset URL” are empty, the Cesium viewer should load asset from the following URL:
https://18.216.24.61/3DTileServer/index.php/asset/<post slug>/tileset.json
Example:
Slug for post https://edd8.construkted.com/downloads/f597bbdf63/ is f597bbdf63

* 2. Fix pagination at URL: https://edd6.construkted.com/gallery/
When page 2 is selected, page 1 results are displayed

* 3. “Geo-Location editor” widget
The lat, long and alt of the model should be displayed to everyone.
The owner should get an extra button in that widget to allow editing of that info,
which brings up the current editing view, along with move widget in 3d view.