Project 30
Once an asset is uploaded to edd wordpress site, the tiler on EC2 server needs to begin
processing the file.
The asset on edd server will be set to “Pending Review”.
File is not public yet.

Once the tiling process is complete, the status of the asset should be changed in
wordpress to “Live