-project 42-2
This text should show the user the status of the processing of a new asset.
Convert the different steps of processing into a percentage and display a percentage.
Use the data from http://tile01.construkted.com:5000/get_active to determine the status
of each post.
The percentage does not have to be perfect. It’s just an indication to the user that the
processing of their file is progressing.
After the file is completely processed and the post status has been changed to
“Published”, status text should be “Complete”.
- Code logic should be; If (post status is “Published”) then echo “Complete”